chore: refine start form styles
